+++ notificare-push/src/main/java/re/notifica/push/NotificarePushCompat.kt
@@ -17,6 +17,7 @@ import re.notifica.push.ktx.INTENT_EXTRA_REMOTE_MESSAGE
 import re.notifica.push.ktx.INTENT_EXTRA_TEXT_RESPONSE
 import re.notifica.push.ktx.INTENT_EXTRA_TOKEN
 import re.notifica.push.ktx.push
+import re.notifica.push.models.NotificareTransport
 
 public object NotificarePushCompat {
 
@@ -81,6 +82,18 @@ public object NotificarePushCompat {
         get() = Notificare.push().hasRemoteNotificationsEnabled
 
     @JvmStatic
+    public val transport: NotificareTransport?
+        get() = Notificare.push().transport
+
+    @JvmStatic
+    public val subscriptionId: String?
+        get() = Notificare.push().subscriptionId
+
+    @JvmStatic
+    public val observableSubscriptionId: LiveData<String?>
+        get() = Notificare.push().observableSubscriptionId
+
+    @JvmStatic
     public val allowedUI: Boolean
         get() = Notificare.push().allowedUI
 
@@ -88,8 +101,8 @@ public object NotificarePushCompat {
     public val observableAllowedUI: LiveData<Boolean> = Notificare.push().observableAllowedUI
 
     @JvmStatic
-    public fun enableRemoteNotifications() {
-        Notificare.push().enableRemoteNotifications()
+    public fun enableRemoteNotifications(callback: NotificareCallback<Unit>) {
+        Notificare.push().enableRemoteNotifications(callback)
     }
 
     @JvmStatic